+++ WeatherWpfApp/Services/IGeoLocationService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace WeatherWpfApp.Services
+{
+    public interface IGeoLocationService
+    {
+        Task<string> GetUserLocationAsync();
+        Task<(double Latitude, double Longitude)> GetUserCoordinatesAsync();
+    }
+}
